presentaion mode

+++ blog/consumers.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import json
 
 from channels import Group
@@ -23,39 +24,51 @@ def disconnect_blog(message):
 
 @channel_session_user
 def save_post(message):
-    text = json.loads(message['text'])['text']
-    title = json.loads(message['text'])['title']
-    author = json.loads(message['text'])['author']
-    method = json.loads(message['text'])['method']
-    id = json.loads(message['text'])['id']
-    requestSender = str(message.user.username)
+    print("***************")
+    click_id = json.loads(message['text']).get('id-click',None)
+    print(click_id)
+    if click_id:
+        # message.reply_channel.send({
+        #     "text": json.dumps({'click': True, 'id': click_id}),
+        # })
+        Group('blog').send({
+            "text": json.dumps({'click': True, 'id': click_id}),
+        })
+    else:
+        print("==========")
+        text = json.loads(message['text'])['text']
+        title = json.loads(message['text'])['title']
+        author = json.loads(message['text'])['author']
+        method = json.loads(message['text'])['method']
+        id = json.loads(message['text'])['id']
+        requestSender = str(message.user.username)
 
-    try:
-        if method == "POST":
-            author = User.objects.get(username=author)
-            Post.objects.create(author=author, title=title, text=text)
-        if method == "UPDATE":
-            postQ = Post.objects.get(id=id)
-            if requestSender == str(postQ.author):
-                postQ.text = text
-                postQ.title = title
-                postQ.save()
-            else:
-                message.reply_channel.send({
-                    "text": json.dumps({'error': "you are not auth this is not your post"}),
-                })
+        try:
+            if method == "POST":
+                author = User.objects.get(username=author)
+                Post.objects.create(author=author, title=title, text=text)
+            if method == "UPDATE":
+                postQ = Post.objects.get(id=id)
+                if requestSender == str(postQ.author):
+                    postQ.text = text
+                    postQ.title = title
+                    postQ.save()
+                else:
+                    message.reply_channel.send({
+                        "text": json.dumps({'error': "you are not auth this is not your post"}),
+                    })
 
-        if method == "DELETE":
-            postQ = Post.objects.get(id=id)
-            if requestSender == str(postQ.author):
-                postQ.delete()
-                Group('blog').send({
-                    "text": json.dumps({'deleted': True, 'id': id}),
-                })
-            else:
-                message.reply_channel.send({
-                    "text": json.dumps({'error': "you are not auth this is not your post"}),
-                })
+            if method == "DELETE":
+                postQ = Post.objects.get(id=id)
+                if requestSender == str(postQ.author):
+                    postQ.delete()
+                    Group('blog').send({
+                        "text": json.dumps({'deleted': True, 'id': id}),
+                    })
+                else:
+                    message.reply_channel.send({
+                        "text": json.dumps({'error': "you are not auth this is not your post"}),
+                    })
 
-    except:
-        return
+        except:
+            return
